# pharma-platform/server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PharmaGuard API
  version: 1.0.0
  description: |
    Pharmaceutical Data Integrity & Compliance Platform.
    FDA 21 CFR Part 11 compliant electronic records, audit trails, 
    e-signatures, batch tracking, and document control APIs.
  contact:
    name: PharmaGuard Engineering
    email: api@pharmaguard.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development
  - url: https://api.pharmaguard.io/api/v1
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for watch agents and external integrations
...
paths:
...
  /batches:
    get:
      tags: [Batches]
      summary: List batch records
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200': { description: Paginated batch list }

    post:
      tags: [Batches]
      summary: Create batch record
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [batchNumber, productName, startDate]
              properties:
                batchNumber: { type: string, example: "BATCH-2026-005" }
                productName: { type: string }
                productCode: { type: string }
                stage: { type: string }
                startDate: { type: string, format: date }
      responses:
        '201': { description: Batch created }
  /batches/{id}:
    get:
      tags: [Batches]
      summary: Get batch with linked files, signatures, audit history
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Full batch record }
  /batches/{id}/release:
    patch:
      tags: [Batches]
      summary: Release or reject batch with e-signature
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [decision, intent]
              properties:
                decision: { type: string, enum: [release, reject] }
                intent: { type: string, example: "All QC tests passed. Batch approved for release." }
      responses:
        '200': { description: Batch released/rejected with e-signature }

  # ─── Users ─────────────────────────────────────────────
...